Change the us Change the us from `uint32_t` to `uint64_t`, because if we use `uint32_t` may cause overflow in some cases.

+++ header/timer.h
@@ -3,6 +3,6 @@
 #include <stdint.h>
 uint64_t get_time(void);
 void delay_ms(uint32_t ms);
-void delay_us(uint32_t us);
+void delay_us(uint64_t us);
 void delay_s(uint32_t s);
 #endif
